perf: refactored app, removed release command
