--- .drone.yml
@@ -1,28 +1,0 @@
-kind: pipeline
-type: docker
-name: release
-
-trigger:
-  event:
-    - push
-    - tag
-
-steps:
-  - name: build and publish docker image
-    image: plugins/docker
-    settings:
-      username:
-        from_secret: docker_username
-      password:
-        from_secret: docker_password
-      repo: docker.pkg.github.com/Semior001/releaseit/releaseit
-      registry: docker.pkg.github.com
-      build_args:
-        - DRONE=${DRONE}
-        - DRONE_TAG=${DRONE_TAG}
-        - DRONE_COMMIT=${DRONE_COMMIT}
-      tags:
-        - ${DRONE_TAG}
-        - latest
-    when:
-      event: tag
